Add console logs and move function
