+++ client/src/components/DraftTool/draftHandler.ts
@@ -111,40 +111,3 @@ export const checkTournamentCode = async (code: string) => {
   }
 };
 
-export const createDraftDBEntry = async (
-  blueName: FormDataEntryValue,
-  redName: FormDataEntryValue,
-  tournamentID: FormDataEntryValue | null
-) => {
-  try {
-    const data = {
-      blueName: blueName,
-      redName: redName,
-      tournamentID: tournamentID,
-    };
-
-    console.log(data);
-
-    // Create the Draft lobby
-    const response = await fetch(
-      "http://localhost:8080/draft/api/createLobby",
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(data),
-      }
-    );
-
-    if (!response.ok) {
-      throw new Error(
-        `Server error: ${response.status} ${response.statusText}`
-      );
-    }
-    const result = await response.json();
-    return result;
-  } catch (err) {
-    console.error("Error in Creating Draft DB Entry: ", err);
-  }
-};
